Save the original request version
When a request is done, it may be tried on different brokers in cas of failing
brokers. Either the api version is enforced when the request was created, or it
is detected against the broker used. However, the version should be detected
again when we swtich broker. It was not the case before, now it is.